pyproject.toml
[tool.poetry]
name = "clover"
version = "0.1.0"
description = "An app to store and serve my personal transactions"
authors = ["martinabeleda <martin.abeleda@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/martinabeleda/clover/"


[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
markdown = "^3.2.1"
psycopg2-binary = "^2.8.4"
pandas = "^1.0.3"
requests = "^2.23.0"
flask-sqlalchemy = "^2.4.1"
flask-restx = "^0.2.0"
flask-marshmallow = "^0.11.0"
marshmallow-sqlalchemy = "^0.22.3"
retrying = "^1.3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
black = "^19.10b0"
pylint = "^2.4.4"
pre-commit = "^2.2.0"
flask-testing = "^0.8.0"
codecov = "^2.0.22"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
